requirements.txt

# ---------------------------------------------
# AI Terminal by Abhi Singh (anodeus)
# Dependency list for pip install -r requirements.txt
# ---------------------------------------------

# Core AI model dependencies
openai
google-generativeai

# Terminal UI enhancements
rich

# System monitoring
psutil

# Web communication
requests

# YAML file support
PyYAML

# HTML parsing / scraping
beautifulsoup4

# ---------------------------------------------
# To install all dependencies, run:
#     pip install -r requirements.txt
# ---------------------------------------------
